Trimmed SQLServerDao class
trimmed down within each method and created 2 new types, 1 that takes a PreparedStatement as a parameter and another the Result Set. this way the switch methods to sort and assign product type is called in each method rather that repeated in each method. has a throws exception which is already in the above block of code so no try/catch needed within